+++ CHANGELOG.md
@@ -4,8 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.1] 2019-08-12
+## [0.5.0] 2019-09-19
 - Allow the `protocol` option in NACL rules to use either string or number
+
+## [0.4.1] 2019-09-18
+### Fixed
+- Add routing to gateway VPC endpoints from both private and secure subnets
+
+## [0.4.0] 2019-09-18
+### Added
+- Add feature to enable gateway VPC endpoints
+
+## [0.3.0] 2019-09-12
+### Added
+- Add feature to enable VPC endpoints
+>>>>>>> 0e72a1809cf262ce640161b346478729871a7a79
 
 ## [0.2.0] 2019-06-12
 - Add `nacl_allow_all_http` and `nacl_allow_all_https` to add http/https egress rules to all NACLs